+++ analyzer/src/test/java/nl/basjes/parse/useragent/analyze/TestYamlBasedExpressions.java
@@ -131,10 +131,16 @@ public class TestYamlBasedExpressions {
         assertTrue(uaa.runTests(false, false));
     }
 
-
     @Test
     public void runIsNullLookupTest() {
         UserAgentAnalyzerTester uaa = createTester("IsNullLookup.yaml");
+        uaa.setVerbose(true);
+        assertTrue(uaa.runTests(false, true));
+    }
+
+    @Test
+    public void runDefaultIfNullTest() {
+        UserAgentAnalyzerTester uaa = createTester("Default-tests.yaml");
         uaa.setVerbose(true);
         assertTrue(uaa.runTests(false, true));
     }